alef_delta_sacco_backend/bot/requirements.txt
aiogram==3.13.1
uvicorn==0.32.0
python-dotenv==1.0.1
